clean up and migrate to ts
